Divide header only files in viewer (for faster compilation), divide them in directories
