+++ threedont/gui/viewer/components/look_at.h
@@ -1,0 +1,23 @@
+#ifndef __LOOKAT_H__
+#define __LOOKAT_H__
+#include "../camera/qt_camera.h"
+#include "../utils/opengl_funcs.h"
+#include <QOpenGLContext>
+#include <QOpenGLShaderProgram>
+#include <QWindow>
+
+class LookAt : protected OpenGLFuncs {
+public:
+  LookAt();
+  void draw(const QtCamera &camera);
+  void setVisible(bool visible);
+  bool getVisible() const;
+
+private:
+  void compileProgram();
+
+  QOpenGLShaderProgram _program;
+  bool _visible;
+};
+
+#endif // __LOOKAT_H__
