+++ threedont/gui/viewer/utils/opengl_funcs.h
@@ -1,0 +1,15 @@
+#ifndef __OPENGLFUNCS_H__
+#define __OPENGLFUNCS_H__
+#include <QOpenGLFunctions>
+#include <QOpenGLFunctions_3_2_Core>
+
+class OpenGLFuncs : public QOpenGLFunctions_3_2_Core {
+  // extends QOpenGLFunctions with some helper and error checking functions
+public:
+  OpenGLFuncs();
+  GLint getBufferSize(GLuint bufferId);
+  void checkError();
+  void printFramebufferStatus();
+};
+
+#endif // __OPENGLFUNCS_H__
